refactor: 重构本地存储IO接口
